requirements.txt
msal
requests
pandas
openpyxl
google-generativeai
httpx
python-dotenv
python-docx
pypdf
